Pagination
Phew!! What a week! I added pagination really quick, but then I realised that the query strings are not working together. Long story short you couldn't sort by oldest todos and go to the next page and vice versa. It was quite an intense week, but finally it is done! F query strings.