Email verification, optional email and more!
This is somewhat a big update. I implemented a functionality, when you enter your email address, to get a mail for verification.

Basically, you need to click on the link provided on the mail to successfully create your account :)

Also I added the option to disable email address, ignoring the need to verify your account. BUT you miss on the feature to reset your password, if needed :)

As a side note, I moved the signalg to their own signals.py file. I did this, because sometimes it can happen to trigger the same signal through importing the class.