Poll Telemetry page, refactoring
This is a rather "big" update. I did a lot of refactoring, and also I added anonymous users telemetry for the Polls. I did it with collecting the ips from anonymous users. They will never be rendered outside the admin panel, so don't worry :)

In the coming days I will write better front end for the statistics page, because right now, it's not the best, but it's there :)